requirements.txt
aiogram
python-dotenv
APScheduler
pymongo
aiohttp
requests
urllib3
aiohttp-cors
numpy
motor
pandas